pyproject.toml
[tool.poetry]
name = "chessengine"
version = "0.1.0"
description = ""
authors = ["Acusick1 <andrewcusick13@outlook.com>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
python-chess = "^1.999"
chess-board = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
